Восстановление работы VGL UI

+++ inc/display2.h
@@ -104,7 +104,7 @@ uint_fast16_t normalize(
 
 uint_fast16_t approximate(
 	const int16_t * points,		// массив позиций входных значений
-	const uint_fast16_t * angles,		// массив позицый выходных значений
+	const int_fast32_t * angles,		// массив позицый выходных значений
 	unsigned n,					// размерность массивов
 	int_fast16_t v				// значение для анализа
 	);
